+++ UnitTests/Sources/MediaProvider/ImageCacheMock.swift
@@ -1,0 +1,42 @@
+//
+// Copyright 2022 New Vector Ltd
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+@testable import ElementX
+import Kingfisher
+import UIKit
+
+class ImageCacheMock: ImageCache {
+    var retrievedImagesInMemory = [String: UIImage]()
+    var retrievedImages = [String: UIImage]()
+    var storedImages = [String: UIImage]()
+
+    override func retrieveImageInMemoryCache(forKey key: String, options: KingfisherParsedOptionsInfo) -> KFCrossPlatformImage? {
+        retrievedImagesInMemory[key]
+    }
+    
+    override func retrieveImage(forKey key: String, options: KingfisherOptionsInfo? = nil, callbackQueue: CallbackQueue = .mainCurrentOrAsync, completionHandler: ((Result<ImageCacheResult, KingfisherError>) -> Void)?) {
+        if let image = retrievedImages[key] {
+            completionHandler?(.success(ImageCacheResult.disk(image)))
+        } else {
+            let error = KingfisherError.cacheError(reason: .imageNotExisting(key: key))
+            completionHandler?(.failure(error))
+        }
+    }
+    
+    // swiftlint:disable:next line_length
+    override func store(_ image: KFCrossPlatformImage, original: Data? = nil, forKey key: String, processorIdentifier identifier: String = "", cacheSerializer serializer: CacheSerializer = DefaultCacheSerializer.default, toDisk: Bool = true, callbackQueue: CallbackQueue = .untouch, completionHandler: ((CacheStoreResult) -> Void)? = nil) {
+        storedImages[key] = image
+    }
+}
