+++ UnitTests/Sources/MediaProvider/MediaProviderTests.swift
@@ -240,97 +240,3 @@ final class MediaProviderTests: XCTestCase {
         return image
     }
 }
-
-private enum MediaProxyMockError: Error {
-    case someError
-}
-
-private class MediaProxyMock: MediaProxyProtocol {
-    var loadMediaThumbnailForSourceData: Data?
-    var loadMediaContentForSourceData: Data?
-    
-    func mediaSourceForURLString(_ urlString: String) -> MediaSourceProxy {
-        MediaSourceProxy(urlString: "test")
-    }
-    
-    func loadMediaContentForSource(_ source: ElementX.MediaSourceProxy) async throws -> Data {
-        if let loadMediaContentForSourceData {
-            return loadMediaContentForSourceData
-        } else {
-            throw MediaProxyMockError.someError
-        }
-    }
-    
-    func loadMediaThumbnailForSource(_ source: ElementX.MediaSourceProxy, width: UInt, height: UInt) async throws -> Data {
-        if let loadMediaThumbnailForSourceData {
-            return loadMediaThumbnailForSourceData
-        } else {
-            throw MediaProxyMockError.someError
-        }
-    }
-}
-
-private enum FileCacheMockError: Error {
-    case someError
-}
-
-private class FileCacheMock: FileCacheProtocol {
-    var fileKey: String?
-    var fileExtension: String?
-    var fileURLToReturn: URL?
-    var storedData: Data?
-    var storedFileExtension: String?
-    var storedFileKey: String?
-    var storeURLToReturn: URL?
-    
-    func file(forKey key: String, fileExtension: String) -> URL? {
-        fileKey = key
-        self.fileExtension = fileExtension
-        return fileURLToReturn
-    }
-    
-    func store(_ data: Data, with fileExtension: String, forKey key: String) throws -> URL {
-        storedData = data
-        storedFileExtension = fileExtension
-        storedFileKey = key
-        if let storeURLToReturn {
-            return storeURLToReturn
-        } else {
-            throw FileCacheMockError.someError
-        }
-    }
-    
-    func remove(forKey key: String, fileExtension: String) throws { }
-    
-    func removeAll() throws { }
-}
-
-private class BackgroundTaskServiceMock: BackgroundTaskServiceProtocol {
-    func startBackgroundTask(withName name: String, isReusable: Bool, expirationHandler: (() -> Void)?) -> ElementX.BackgroundTaskProtocol? {
-        nil
-    }
-}
-
-private class ImageCacheMock: ImageCache {
-    var retrievedImagesInMemory = [String: UIImage]()
-    var retrievedImages = [String: UIImage]()
-    var storedImages = [String: UIImage]()
-
-    override func retrieveImageInMemoryCache(forKey key: String, options: KingfisherParsedOptionsInfo) -> KFCrossPlatformImage? {
-        retrievedImagesInMemory[key]
-    }
-    
-    override func retrieveImage(forKey key: String, options: KingfisherOptionsInfo? = nil, callbackQueue: CallbackQueue = .mainCurrentOrAsync, completionHandler: ((Result<ImageCacheResult, KingfisherError>) -> Void)?) {
-        if let image = retrievedImages[key] {
-            completionHandler?(.success(ImageCacheResult.disk(image)))
-        } else {
-            let error = KingfisherError.cacheError(reason: .imageNotExisting(key: key))
-            completionHandler?(.failure(error))
-        }
-    }
-    
-    // swiftlint:disable:next line_length
-    override func store(_ image: KFCrossPlatformImage, original: Data? = nil, forKey key: String, processorIdentifier identifier: String = "", cacheSerializer serializer: CacheSerializer = DefaultCacheSerializer.default, toDisk: Bool = true, callbackQueue: CallbackQueue = .untouch, completionHandler: ((CacheStoreResult) -> Void)? = nil) {
-        storedImages[key] = image
-    }
-}
